refactor(datasets): render Upload CSV as a drawer, not a centered modal
The Upload CSV surface was registered as a drawer (drawerRegistry.uploadCSV)
but rendered a centered Dialog. Swap the primitive to Drawer (size xl, matching
the AddOrEditDatasetDrawer it hands off to, so the upload→edit handoff keeps the
same width and header style) and rename UploadCSVModal → UploadCSVDrawer so the
component name matches what it renders.
